First draft of language-generic types documentation.
This is in preparation for a second CL which introduces a whole section for Rust. I split this out first since it's a big chunk on its own, and stands alone.

---

Types are interesting in that they are inherently bidirectional, to a large extent: `i32` **is** `int32_t`.

Before this CL, we had a C++ doc that said, basically, "`int32_t` becomes `i32`". That's fine. But with Rust support on the way, we're faced with the prospect of writing a second doc, which says "`i32` becomes `int32_t`", and so on, down the line. Unlike language features, types have an inherent bidirectionality to them, where it makes sense to make the documentation cross-language.

Our functional docs will have three major sections, then:

* types
* C++ language features supported by Crubit (including how to define new types, in C++)
* Rust language features supported by Crubit (including how to define new types, in Rust)

I _think_ this makes sense.

---

Note that this is _partially_ undoing the original work of separating out the docs: originally e.g. C++ enums and Rust enums got the same page. I still think that was the right split to do: there's not so much in common, from a user or documentation POV. But for types, there's a lot that's in common, and it makes sense to deduplicate. Though it's a much easier sell for the `primitive.md` page than the `pointer.md` page.

PiperOrigin-RevId: 680710503
Change-Id: I983ae9da4ece5fc21e276070527913df3e3877d1