Allow non-trivially destructible Rust types to be passed by value.
This is the same thing as how we deal with them being _returned_ by value, but in the opposite direction: just as Rust used `MaybeUninit` to prevent destruction of the Rust side, C++ callers must use `NoDestructor` to prevent destruction on the C++ side.

In fact, this caught me by surprise: without testing, I'd thought we did it this way, oops.

---

Consider the following Rust struct:

```rust
#[derive(Default)]
pub struct Foo {
  pub x: i32
}

impl Drop for Foo {
  fn drop(&mut self) {}
}
```

This is movable in C++, but not _trivially_ movable.

Crubit will (before this CL) not generate bindings for the following function:

```rust
pub fn foo(_: Foo) {}
```

With the justification "Only trivially-movable and trivially-destructible types may be passed by value over the FFI boundary"

This restriction is artificial / can be lifted by changing how we generate thunks. The problem is that in the thunk, we pass non-abi-compatible objects by pointer, and copy the bytes to move out of that pointer. What we _need_ to do, instead, is perform a C++ move -- leaving behind a **moved-from** value, which can be safely destroyed.

For example, instead of this C++ code:

```c++
void foo(Foo __param0) {
  return thunk(&__param0); // BAD: __param0 will be destroyed after copy
}
```

We can generate maybe something like this:

```c++
void foo(Foo __param0) {
  return thunk(absl::NoDestructor(std::move(__param0)).get());
}
```

This lets us leave behind a valid __param0 (a moved-from __param0), which works with any move-constructible `Foo`. And the `NoDestructor` will not invoke the destructor when it cleans up storage, meaning the bits can be copied out in a Rust move.

Strictly speaking, this performs one more copy than necessary: ideally we'd move directly from the parameter to the final location, not to a temporary holding location. That would require running the logic of `std::move` from Rust instead of C++, and is trickier / requires more code duplication. For now, I decided the easiest way is to just do this.

(Originally, I was going to file a bug, and this was the bug description! But when I realized there was this very-very-simple approach, I went with that.)

PiperOrigin-RevId: 681655143
Change-Id: I5e8c3b0f335a745dc67d63a3b045b290a3367554