Add slice_ref to _cc_deps_for_bindings
h/t dmitrig@ for finding the missing place.

Adjusts the fission_test to check for expected PIC files being a subset of the actual ones, instead of equal, because the slice_ref dep brings more.

This should make it possible to find the slice_ref.h header when referenced in the generated header files as:

```
#include <crubit/support/rs_std/slice_ref.h>
```

PiperOrigin-RevId: 683514337
Change-Id: I10eb6145a1b745faa88f6b9a71e00ac01a4b62d5