Update type alias and use declaration docs.
I switched this to the Rust technical term "use declaration", which is what it's called e.g. here: https://doc.rust-lang.org/reference/items/use-declarations.html.

I also tried to characterize the behavior a bit more, but IDK how successfully.

Finally, I harmonized the wording of the docs as a whole, and used the copy-pastable example code in the examples sections.

PiperOrigin-RevId: 684119407
Change-Id: Iefa1fe660e9d8df0ac6431de213c7da3c910ded6