Ready-for-release version of bidirectional Crubit documentation.
This removes all "experimental", "not yet released", etc. markers, and finally documenting cc_bindings_from_rs in the `README.md`.

I plan on submitting this immediately before the announcement email is sent out.

Docs aren't ever "final", but this is the last step in the **initial** documentation process that I wanted to go through, and the rest is just keeping them up to date, documenting new features, etc.

PiperOrigin-RevId: 684210257
Change-Id: I4e2c9ce89b73be12a69abbb873a41bdf46f93c66