[nullability] Loosen assertion to allow nullability attributes in type aliases.
When nullability type aliases are defined with nullability attributes, i.e. like
this:

```
template <class T>
using Nullable = T _Nullable;
```

then `NullabilityWalker::sawNullability()` will be called twice, once for the
type alias and once for the attributed type contained within it. This would
cause the existing assertion in `sawNullability()` to fail in the case where
the nullability is `_Null_unspecified`.

This change loosens the assertion to allow one unspecified nullability to wrap
another. The type aliases in `GetTypeNullabilityTest` are changed to add
nullability attributes, causing tests to fail without the fix.

I have had to delete one specific case from the `Overrides` test; this is
because Clang does not allow conflicting nullability attributes on the same
pointer. Note that this only applies if the nullability attributes are specified
directly; if they are specified through type aliases (as will usually be the
case), they ambiguity is resolved, and in the same way as `NullabilityWalker`
resolves the ambiguity, i.e. the outermost nullability "wins". See this
[godbolt](https://godbolt.org/z/3Gh3vzfqG) for a demonstration of this.

PiperOrigin-RevId: 684317108
Change-Id: Ib6885240d5e631ea662fdb2d96c3f1ac13288433