Remove redundant diagnostic of smart pointer arrow operator.
We currently explicitly warn on `operator->` for smart pointers. But, they don't actually dereference, they just return a pointer, which is in turn dereferenced by the the MemberExpr. So, we've been generating 2 diagnostics in this case. That used to be hidden, because they both covered the same range and were deduped, but then we changed the range for normal arrow, and now they produce duplicate diagnostics.

PiperOrigin-RevId: 684920630
Change-Id: Ic688d2d048c40a7f35ad8207022d5123038dd56b