Set `defining_target` for explicit template specializations.
There was an inconsistency here: while it's true that, in principle, since there's only one target involved here (the template specialization is defined in the target that owns it), the existence or nonexistence of `defining_target` is used elsewhere to determine whether a type is a template type. So we still need to populate the `defining_target` field, in order to e.g. filter it out correctly in `:supported` targets.

PiperOrigin-RevId: 684974881
Change-Id: I8bad2f6179d9cd36d7c4b0a57fa41e13512f8f84