Make component of cc_std not dependent on Crubit.
Crubit automatically include `cc_std` as deps for its generated bindings. If we include the generated bindings of some additional headers as part of the `cc_std`, a loop dependency is introduced. Instead, we directly include these headers as an internal part of the `cc_std`.

I encountered this when I try to include `Vector` as part of `cc_std`.

PiperOrigin-RevId: 685883090
Change-Id: Ida5b3f2b77ac7d3cbe9f68a5b93830f1abb2576c