[nullability] Find underlying pointer type correctly when no type alias present.
Previously, we didn't always correctly find the underlying raw pointer type for
a user-defined smart pointer when that smart pointer didn't define either the
`pointer` or `element_type` type aliases. The intention was that we would then
deduce the underlying raw pointer type from the template argument (if
applicable).

Ironically, we _did_ do this correctly in the case where we didn't have a
definition for the class template specialization, but not if we _did_ have a
definition.

This change fixes the false negative in smart_pointers_diagnosis.cc, and it also
adds a direct test for `underlyingRawPointerType()` (that fails without the fix)
in type_nullability_test.cc.

PiperOrigin-RevId: 688845656
Change-Id: I0d175cbd63ac8dd7a3e361a122619c1b02513d29