Put additional sources in the canonical namespace so that the source can be correctly located.
If the namespace `ns` is reopened, Crubit generates intermediate modules with number suffixes `ns_0`, `ns_1` except for the last reopened one, which gets name `ns`. We should export the additional sources to the canonical namespace, because otherwise the compiler cannot locate the source file in path like `ns_1/some_path`.

PiperOrigin-RevId: 690793435
Change-Id: I91d98b63b5959e3b80b092d0ae3f898edc6c9f02