[nullability] Fix a crash in smart pointer assignment
When the LHS is a smart pointer, but the RHS is not (and not nullptr), we can't getPointerValueFromSmartPointer(RHS). Instead, set the LHS nullability to Unspecified to model not knowing the nullability of the RHS.
We could handle more cases, like when the RHS is a lambda.

This case happened to come up once we started handling smart pointers without `pointer` or `element_type` type alias.

PiperOrigin-RevId: 691771199
Change-Id: Id0beed0f324f0fed619aaecbddfd3ea86842a912