Allow feature allocator_api.
`cpp_std_allocator` requires `allocator_api`. We need this allocator in order to be able to have zero-cost sharing of allocations between C++ and Rust in a cross-platform way. Currently only `cc_std::Vector` is using it.

As we will move cpp_std_allocator, which requires allocator_api, to cc_std. As cc_std is by default a dependency of all the bindings, we turn on `allocator_api` by default.

PiperOrigin-RevId: 691848914
Change-Id: I8e95d57332b282351102f408c8159438cc1a5bef