Fix a crash when analyzing smart pointers with operators in a base class.
The AST matchers appear to ignore implicit casts, so if operator* is defined in a BaseClass, we still consider the receiver type ChildClass in the case of `ChildClass instance = ...; *instance`, instead of BaseClass.

If only ChildClass is marked as nullability compatible and BaseClass is not, then we currently crash due to a mismatch.

Make the transfer functions also ignore implicit casts to match.

PiperOrigin-RevId: 693017689
Change-Id: I7d8328ab380c96d864aca8f3bc905e5a6e457f97