Fix a crash when diagnosing operator-> on smart pointers when the guessed pointer type isn't correct
Currently, we bail out during transfer if the guessed pointer type is incorrect in operator->, get(), release() (e.g. if we guess T** instead of T*).

However, when diagnosis attempts to look up the nullability state of the Call's result, which is often a raw pointer type, it isn't found.

Instead, try to initialize the Call's return value to something.

PiperOrigin-RevId: 693344553
Change-Id: Ibec54d2cf81c9972623e94f415cfc383b2be98a0