Add release-safety flags for cc_bindings_from_rs.
* A default-features flag, so that the set of features enabled by default can be changed by a flag flip (not a binary release)
* An optout crate-disabled-feature flag, so that features which are rolled out by default can be rolled _back_.

This is unlike rs_bindings_for_cc (for now), because cc_bindings_from_rs actually _has_ default-enabled features! (Whereas by default, right now, rs_bindings_from_cc is disabled). So we need a little bit of extra infrastructure to make it possible to safely roll out / roll back dangerous or backwards-incompatible changes.

P.S.: the other way of doing this is to do it entirely in bazel, but that explodes the size of the command line, so I like the efficiency of this. :)))

PiperOrigin-RevId: 693867266
Change-Id: I64749f1fcab756cbe610fd499e9eb1e53c20c520