This CL refactors the copy-paste attribute parsing in `crubit_attr::get` into a loop, which should be much more extensible for adding future attributes. In order to accomplish this, we add dynamic setter/getter methods to CrubitAttr (rather than having to statically reference the field name, which requires an unrolled loop).
This CL also updates the name and documentation for the `CrubitAttr` type to make it clear that it represents a collection of attributes, rather than a single attribute. (Also slightly shuffled and updated some documentation for conciseness and clarity.) This CL similarly updates the name of the `crubit_attr::get` to help indicate that it returns *all* the attributes of the given definition.

PiperOrigin-RevId: 693960761
Change-Id: If6ece7fa0792acfc2c4c027b0c74e36099ac8033