Fix a crash for missing pointer nullability state when smart pointer transfer function bails out
We could encounter an "unusual" `operator*` which returns a raw pointer instead of a reference to T, and diagnoseArrow expects to be able to find the state of the raw pointer.

Alternatively, (haven't tried yet) we could change the matcher to be stricter and only use this transfer function when has the expected return type.

PiperOrigin-RevId: 694196743
Change-Id: I0a68da6e830842be5c6a2ff9906b85d839c40531