Remove the slot field of a SlotRange.
This allows the same SlotRange to represent a single source range that supplies the type for two different nullability slots (aka two different positions in potentially two different nullability vectors). This is a rare situation, but can occur with macros that accept type arguments and use the same type argument in multiple places.

PiperOrigin-RevId: 695431990
Change-Id: I1a3b041c87f0eb603004db9d3c8ca81ea5b0dec3