[nullability] Remove superfluous check from `underlyingRawPointerType()`.
A recent change made it so that we check for a template argument on the smart
pointer base class in addition to the derived class. However, we should never
have been looking at the derived class in the first place; we know the base
class is the actual smart pointer class, and so that's the only place we should
be looking at the template argument.

PiperOrigin-RevId: 695721616
Change-Id: Id90ae91aabffed05d51bad86ac3bee78bc8066f1