More precisely collect evidence from the taking of references to pointers.
When a reference to a mutable pointer is taken, their type nullabilities must match, not just the nullability of the decl's current value, to ensure that the underlying value is not made inappropriately null by assignment through the reference.

When a reference declared as referring to a const Nonnull pointer is taken, the left and right side type nullabilities must also match, to ensure the underlying value is not made inappropriately null directly or through a different reference that is declared as pointing to a Nullable pointer.

Also add some aspirational test cases for assignments to pointers to mutable pointers and some cases of copy assignment through references.

PiperOrigin-RevId: 700398379
Change-Id: I5eba1c4fffcdc8de589b3d5d59d9430ddd42700b