Collect evidence from lambda captures with initialization.
These nodes aren't seen in the body of the `operator()`, but when they are seen as VarDecls in the AST, their DeclContext is the `operator()`, so we were not considering it to be an available initializer to collect evidence from. We can't annotate them, since the types are inferred, but as seen in a test case, their inferred nullability can propagate to other variables that we can annotate.

PiperOrigin-RevId: 702042519
Change-Id: Ibb3d8be1cf077025125f686a8a816e2172557711