Bridge pointer-like ADTs without conversion functions #crubit #cc_bindings_from_rs
This change optimizes a special case when bridging Rust to C++ types, where a Rust struct is effectively a pointer i.e. it contains a single field that is a pointer. If so we transmute the memory to a C++ pointer.

Given the below type definition:

#[__crubit::annotate(
  cpp_type = "const CppType*",
  cpp_type_include = "ns/cpp_type.h",
)]
#[repr(transparent)]
struct PtrLike {
  data: NonNull<c_void>,
}

If Crubit encounters this type in a function argument or return value, the corresponding C++ bindings will use `const CppTye*` instead. Crubit verifies that `PtrLike` has the same in-memory representation as a C pointer and if so bridges the by performing bitwise copy (transmute).  We require the type to be annotated with repr(transparent) to ensure that the code continues to be correct when run under sanitizers, which may add additional fields.

PiperOrigin-RevId: 702320579
Change-Id: I9241448d603fd8283b77630fdb2303de79cfd016