De-duplicate extern "C" declarations #crubit #cc_bindings_from_rs
Before this change if a bridged type appeared in multiple arguments or return types, the generated bindings would fail to compile because for type conversion an extern "C" declaration would be generated. rustc allows only a single extern "C" declaration per symbol. The declarations are generated at the callside of the type converter functions and therefore they don't have a global view that would allow them to declare each function only once.

This change makes it so that all extern "C" declarations are generated in a central place and thereby they can be deduplicated. We use a BTreeSet to deduplicate by symbol name and achieve deterministic ordering.

PiperOrigin-RevId: 702331690
Change-Id: Ia11e3aecc6f7b5893d5c2c3719db7d9448da0e29