Correctly collect TypeNullabilityLocs and eligible ranges for qualified elaborated types.
There were 2 overlapping bugs: that we didn't include the type elaboration in the TypeNullabilityLocs and that we would have attempted to include preceding CVR qualifiers in ranges for smart pointers. Both are fixed in this change, so we now produce correct ranges for qualified and/or elaborated smart pointer types.

PiperOrigin-RevId: 704703459
Change-Id: Ic6e9f50f9c75339d299dd030d5b4e10705d1f927