Fix for Crubit generating invalid Rust syntax
Crubit generates invalid Rust syntax if a function has at least one argument whose type is a bridged type and a return type that is not bridged.

The code would take the form `__ret_ptr.write(let foo = ...);` and miss a block expression `{ }`. This change unconditionally wraps the arguments of `__ret_ptr.write()` in a block expression: `__ret_ptr.write({ ... })`.

PiperOrigin-RevId: 707006513
Change-Id: I152a8ce7abc6d11a64551f80d7b01b79cd81bbfe