Add &dyn BindingsGenerator as input to RsTypeKind::is_unsafe
The parameter is unused right now, but allows for upcoming unsafety changes which will depend on querying the db to ask if an RsTypeKind is unsafe. This is an alternative approach to trying to cache an `is_unsafe_type` bool within RsTypeKind::Record, which runs into problems with cycles in the db and code duplication.

PiperOrigin-RevId: 716331961
Change-Id: Id283d456521e980b86b983a5d20573c570e0aa62