Support for more smart-pointer-like class caching (classes with * and -> returning the expected type)
Use the smart pointer caching utilities (matcher and transfer function) from clang::dataflow (https://github.com/llvm/llvm-project/pull/120249/) to handle some more accessor caching cases.

This is similar to the `transferValue_OptionalOperatorArrowCall` in that it can treat the non-const overload as const, but it also allows mixing the operator * vs -> vs value() between the check and the deref. It also handles more classes like absl::StatusOr.

Change the test stub in `OptionalOperatorArrowCall` a bit, because the heuristic for matching smart pointers classes requires `*` and `->` (which the real version of std::optional would have).

PiperOrigin-RevId: 718973436
Change-Id: I1bdf3061ce5658e38e7f6b0225bdbed512ccc301