Document a bit more of our template body verification support with tests
In particular, have a case where "target" itself is templated and has a violation. This requires skipping isTemplated matches in the test framework and checking that we at least saw one instantiation (not isTemplated).

It seems that template bodies that are unsafe independent of the template type params can be flagged (e.g., if you just had a T *x = nullptr; *x;). However, cases that depend on the template type param are not flagged.

Existing tests seem to only check uses in "target" e.g., the return type
or that args match params (vs target itself being a templated function).

PiperOrigin-RevId: 720679648
Change-Id: I660da57854797eea8e249b0aed34cb96859c9f19