Make `RsTypeKind::new_record()` always return a valid RsTypeKind.
Before this change, `new_record` would return a BridgeType when the type is a bridged template specialization, but the plain Record if the bridging was instead by an attribute. This is makes it hard to treat bridging in one consistent way. This led to checks like:

```rust
if record_rs_type_kind.is_bridge_type() || record.bridge_type_info.is_some() {
```

Moreover, it can lead to bugs: if we call `new_record` directly, in this old state of the world, the returned Record could be one that doesn't even exist, and has an incorrect name, etc.. That's dangerous.

After this change, all bridged types (whether they are special-cased templates, or annotated) are bridged in every place an RsTypeKind is created, including when calling `new_record` directly.

PiperOrigin-RevId: 723246641
Change-Id: I63733739e4ed2a1403d0190761eb9cd4eaf15892