Disable the generic bindings for `unique_ptr<T>` and `vector<T>` if T is bridged.
unique_ptr<T> and vector<T> are specially implemented to reinterpret the type as a Rust generic -- but this only works if the generic parameter, T, is actually layout-compatible and has a compatible destructor which does the right thing, and so on. Bridging, by definition, defeats this. But both bridging and this reinterpretation were added separately, and neither knows about the other. So we have to add that here. :)

The added tests not only fail, but actually segfault before this CL. Very bad.

PiperOrigin-RevId: 723405511
Change-Id: Icc989f4581adb2e56f975a9265965adeacf30020