run_compiler_test_support: support transition in internal rustc API
New in this change is that we're using a macro from the rustversion
crate to accomplish the job, which makes life much easier.

Sadly, our toolchains had been marked as -dev before, so we can't use the before() and since() macros in this fix. In the future we should be able to do that though!

PiperOrigin-RevId: 724146016
Change-Id: I02f1bfddf0c0f5d85030db07409eed01b6f7b844