check_include_directories.sh: Query Clang for list of system include dirs.
So far, we have simply checked the list of `-isystem` flags on the command line.
However, I plan to make a change soon that will make Crubit find the builtin
header files in the same way that the Clang compiler itself does (namely by
finding the resource directory using a relative path from the compiler driver,
or, in the case of Crubit, rs_bindings_from_cc_driver). When the builtin
header files are located in this way, they don't appear on the command line as a
`-isystem` flag, but they do appear in the list of system include directories
that Clang outputs when run with the `-v` flag.

A side benefit, IMO, is that the test becomes shorter and easier to read.

PiperOrigin-RevId: 725083581
Change-Id: I7f81d9352b8d856508ec4c99ba9823a23284334a