Add cc_std::std::raw_string_view
This is the first step in implementing an API change to introduce lifetimes for string_view.

For now, cc_std::std::raw_string_view is being made an alias for cc_std::std::string_view. This enables migrating usage of the latter to the former. In follow-up CLs we will do the migration and then actually replace string_view with raw_string_view. That will allow us to redefine string_view as a type with a lifetime.

PiperOrigin-RevId: 725399266
Change-Id: I76080acf07e03b326ac4c7ea1f1da4fb4a2cad4a