Move work out of the generate_bindings crate
After this CL, generate_bindings finally takes <1s to compile.
This makes iterations on the library and tests much faster.

Note the change to the `ir::deserialize_ir` from `<R: Read>`
to `&[u8]`. This was critical, as it moved the generation of
the deserialization code out of the `generate_bindings` crate
and into the `ir` crate.

PiperOrigin-RevId: 725636421
Change-Id: I70d9cdc8095ec8eacc8c311f74dc71405a0c3551