Exclude decls which are templates or are in templates from range collection.
Dependent types can't be walked by the NullabilityWalker, so we largely don't end up even trying to collect ranges in these cases, but we can still hit some sharp corners of hidden dependent types if we try.

This fixes crashes seen when running on real code.

Ranges for instantiated templates will still be collected when walking the instantiations.

PiperOrigin-RevId: 725739375
Change-Id: Iabe3c5e0980592ad1bcffd4d52e90c20aeb97c2e