Find builtin headers relative to rs_bindings_from_cc_driver script
Part 1 of 2: This change introduces a new flag `--driver_path` that can be used
to pass the path of the driver script, but it doesn't yet remove the code in
generate_bindings.bzl that passes the path to the builtin headers as an
`-isystem` flag. I'm separating the changes out in this way to show that it's
harmless to release a version where both behaviors coexist (i.e. where we can
find the builtin headers both through the `-isystem` flag as well as relative to
the driver script).

PiperOrigin-RevId: 726041985
Change-Id: I4627f9580b8c0d4ae2681f06b7d28da87be8d61b