Use the linear-time glob algorithm for our quoted patterns in the fast path.
This speeds up our tests by somewhere in the order of 10x:

| Test                        | Before | After |
|-----------------------------|--------|-------|
| `ir_from_cc_test-arm_linux` | 23.8s  | 2.9s  |
| `ir_matchers_test`          | 23.7s  | 0.7s  |

Originally, I had the idea of replacing the recursive approach. But the error messages turned out to be atrocious. In part because this is because this approach discards all the state you'd need for a good error message, but also because I brute-forced the "inner" search / it is a flat search done repeatedly, and not a tree search.

The fast path is potentially a little bit more featureful I think, which is not good. (E.g. it adds efficient backtracking in general). This means that if you lean on something provided in the fast path, but fail to match, the error messages might be confusing and refer to things that only work in the fast path and not the slow path. But in practice I think this is not so bad and worth it for the performance wins.

---

This CL is like the third time I've implemented linear time globs, and like the twelfth time I've implemented one of Russ Cox's blog posts. Damn it all. And, yet again, I got confused by -- do you know how subtle it is that all but one checks for `nx` use `< len(name)`, but one, very crucially, uses `<= len(name)`? That bites me every time!

(and `nx`, what is that, a variable name for ants?)

PiperOrigin-RevId: 726233643
Change-Id: Ic07b78bd827aaaebeff8e7894fd74b3b1e686181