Document safety/soundness requirements for Crubit, including allowances for safe `Clone`.
I've redrafted a bit, but it could still use some polish.

The strongest motivation here is to make the default constructor Just Work as `Default::default()`, and to have a rational / documented basis for this, which also extends to explaining why we think we can call the destructor from `Drop`, or even call the copy-assignment operator from `Clone::clone_from`. This is, I think, just barely squeaking it through: "exclusivity law" violations are forbidden by (2) (any references you keep around which would cause mutable aliasing would **also** cause use-after-free if you do a Rust-move), and validity is guaranteed otherwise by (1), which is itself a nearly overpoweringly strong assumption.

This page is modeled a bit on https://duchess-rs.github.io/duchess/safety.html, but builds specifically on top of Łukasz's work documenting the reproducible builds subset of the requirements.

PiperOrigin-RevId: 731395847
Change-Id: I1a4a72081ad9e917336ddeb007090294f838f3d3