Don't generate struct fields if the field type is a bridged type.
In general a bridged type is not layout-compatible with the original type. It can be in some cases but we don't have this information available in Crubit at the moment.

This change makes it so that no typed struct field is generated if the field type is a bridged type. Before this change we would get compile errors for fields of bridged types that are not layout-compatible.

PiperOrigin-RevId: 733596963
Change-Id: I4b53eb0757a9f721d83a687238c0458650772b7a