Change LocFilter choice from a boolean to an enum
We will be adding more cases as a next step to allow finer grained filtering of test files (skip just the "main" file that is the test, but NOT elsewhere -- either the associated header or other dependency headers).

Test files may instantiate templates in other headers, and we don't want to filter out the template bodies and instead analyze them.

PiperOrigin-RevId: 738856438
Change-Id: I5170556d112184ab933bc84b6af4f2f175b215e1