Add a proc macro for user-provided Crubit attributes
This change introduces a proc-macro which generates Crubit attributes as
doc comments, allowing future changes to replace the __crubit tool
attributes.

In contrast to the current tool attributes solution, this approach:

* Offers clear, spanned compile-time errors in response to misused attributes.
* Provides a single library-defined location for API docs.
* Allows users to avoid direct dependencies on an unstable feature.

This opens the door to providing more widely-used Crubit annotations.

Note that attributes are passed through doc comment syntax in order to allow
for easy pass-through on the stable Rust compiler. This approach is used by
other bindings-generation libraries.

PiperOrigin-RevId: 740504189
Change-Id: Iad62efb237b5708a2032a88fb89b9412332a8484