Document (and implement) a future-compatible way to implement possibly-conflicting additional Rust sources.
This CL does two things:

1) document that you should use CRUBIT_RUST_NAME to rename things if you want to use that name in additional rust sources. For example, if you want to define Clone, rename the copy constructor.
2) actually implement the fix that makes (1) work.

The change to make CRUBIT_RUST_NAME work for trait implementations is not currently useful: the bindings are not _callable_ in any useful way, even if you are using the experimental Crubit features and so on. That's OK: we can fix that up later so that they become factory functions. Producing uncallable bindings (which compile) still suffices for the  use case here.

I filed b/402478920 to make CRUBIT_RUST_NAME "actually work" here.

PiperOrigin-RevId: 740937114
Change-Id: Id1e929ca441d19c449e5c5f787809efb5a51aa2a