Handle nullptr literals in inference's make_unique special-casing
With make_unique we will miss the cast from arg -> param type, and are only left with type `nullptr_t`. nullptr_t don't have PointerValues, so we change some code to operate on the PointerNullability, which we can make up given either a PointerValue or knowledge that the expression is a nullptr literal.

PiperOrigin-RevId: 740963513
Change-Id: Ic28dfc4bc6c88d4ff7c612fe5c02071ed840f83c