Consume proc macro-generated doc attributes for user-provided Crubit attributes
This change moves away from the unstable tool interface onto
a procedural macro. This has several advantages:

* Offers clear, spanned compile-time errors in response to misused attributes.
* Provides a single library-defined location for API docs.
* Allows users to avoid direct dependencies on an unstable feature.

This opens the door to providing more widely-used Crubit annotations.

Note that attributes are passed through doc comment syntax in order to allow
for easy pass-through on the stable Rust compiler. This approach is used by
other bindings-generation libraries.

Also, targets which are not currently able to directly depend on the proc
macro are able to express their usage through doc attributes directly.
This should be cleaned up in future CLs as more of these targets are
allowed to use the proc macro directly.

PiperOrigin-RevId: 742325500
Change-Id: I5544baad497413a1b0377d887b4bddc47e6e709a