Replace RsTypeKind::Option with flags in Reference and FuncPtr
It was previously confusing what `RsTypeKind::Option` meant. Was it talking about things that can be nullable (refs/fnptrs)? Or could the inside be any type? It turns out, it was the former, and that the inner type of `RsTypeKind::Option` was only intended to be a reference or a fnptr. This CL cleans up the code by removing the `RsTypeKind::Option` variant and instead adding an `option` bool to `RsTypeKind::Reference` and `RsTypeKind::FuncPtr` instead, making impossible states unrepresentable.

PiperOrigin-RevId: 744050431
Change-Id: Ibbd8d1d005137c97a8425cef299c6819506f3e47