Clean up arguments in ConvertQualType
This CL does two things:
* First, it fixes a bug where in the newly expanded CcType, I forgot to add the check that turns "this" pointers of methods into rvalue refs if they are ref qualified (e.g. `void f() &&`) and have a lifetime.
* Second, this helped me get a deeper understand of the `ref_qualifier_kind` and `nullable` arguments that were getting threaded through ConvertQualType and friends. I realized that `ref_qualifier_kind` is only ever used for converting the QualType of the "this" pointer for methods. I was able to take it out of all those functions, while still maintaining behavior by just adding a small adjustment after ConvertQualType was called on the "this" pointer type in functions.cc. This is great because it reduces the number of arguments for a bunch of callers, where they were all passing in `std::nullopt` anyways.

I also learned that `nullable` is only false for the "this" pointer from methods. I tried to remove that too, but ended up hitting strange errors with ItemIds not being found in the IR in some tests. Since we'll probably want to add nonnull pointer annotation support in the future, I left that in and added a comment for future readers about who is using `nullable=false` today.

PiperOrigin-RevId: 744059483
Change-Id: Iece978b568af3405e40e22c24f4a726c71d82f08