Create RsTypeKind from CcType instead of RsType.
The problem with RsType was that it was too weakly typed, as evident from the diffs in rs_type_kind.rs. This was causing problems because we need to have RsTypeKind understand C++ std types that have bridge logic, and RsType was becoming difficult to work with because of its usage of stringly typed variants.

Another question in the code base was "what is the purpose of having RsTypeKind and RsType?" And there wasn't really an answer for that- RsType was just an intermediate type for creating RsTypeKind. So instead of improving RsType and leaving CcType as it was (it didn't really have any issues), I decided to instead improve CcType to use more descriptive types, and have it create RsTypeKinds directly without having to go through an RsType. That is what this CL does, but much of the work of making CcType better occurred prior (https://github.com/google/crubit/commit/1c526cbb15fbdab3afb92aacf87ba606761125a0, https://github.com/google/crubit/commit/32bb5b5f5f6c42361d25269a72c4f65a784b454a, unknown commit, unknown commit).

There are several next steps:
* Remove RsType completely. We still create RsTypes, and use them for one or two things that's not creating RsTypeKind. These usages should be migrated to CcType, and RsType can be deleted, and MappedType can be inlined to just CcType. This isn't necessarily a priority, but it's low hanging fruit for lots of code deletion :)
* Actually use the new custom bridge type information. This CL threads the data about if a type is a std bridge type right up to RsTypeKind creation in rs_type_kind.rs, but doesn't do anything with it yet. I need to follow up and use this information to get bridging working for them too.

PiperOrigin-RevId: 744857213
Change-Id: I2a8ce4cdf8de3fdd121fcde0637f949632e770a3