Fix custom bridge types with template types not getting bridge data
The bug was that when adding bridge data in `cxx_record.cc`, I would check if a type was a template specialization. If it was, I would only check if it was a builtin type. If it wasn't, I would then check for the user annotation. But if it was a template specialization with a user defined bridge annotation, then that would never get seen. This CL fixes that.

PiperOrigin-RevId: 745209111
Change-Id: Ifa490966ddb4b209d2379023fd7a76ae492f24da