Replace PrimitiveType with Primitive
This is some initial setup for b/409794384, which will require types that use the `TransmuteAbi` to be able to spell themselves in Rust and C++. For example, bridging a type that wraps int will need to spell out `TransmuteAbi<int>` in the C++ codegen, so it's important that we don't lose that information completely by going to RsTypeKind. Also, PrimitiveType and Primitive had a lot of overlap so I don't feel too bad about this.

PiperOrigin-RevId: 747483426
Change-Id: Iaf64baaa648d0623ed4300c2d35bf58cf0697b1f