Rename `CcCast::cc_cast` to `CppCast::cpp_cast`.
This has been on the to-do list before we get `forward_declare.rs` into a supportable state. Outside of the build system, we should be using the abbreviation `cpp` for everything -- some unscientific polling, and the very name of the official website (isocpp.org) and conference (CppCon), point to "cpp" as the semi-official identifier for C++, rather than `cc` (a bazelism) or `cxx` (mostly just the crate!)

Since `CcCast::cc_cast` has zero uses outside of tests at the moment, this is a pretty easy change to make now, and we should make it now before it becomes difficult.

PiperOrigin-RevId: 747498925
Change-Id: I14d2dec8f567f7777c9dc16f53cd194a47dafed2