Resugar member operator calls with the "base" object nullability to refine
member operator call ASTs have a different shape from normal method calls where (1) "base" is a sibling of the callee, and not a child through a MemberExpr.
(2) "callee" isn't a MemberExpr (there is no MemberExpr in the AST)

So, before, the resugaring in transferType_MemberExpr isn't done for operator calls.  Do something similar to transferType_MemberExpr, but at the point of the operator call, where the base and the callee are both accessible as children.

That seems enough to refine the return nullability, but may be too late to affect the full callee signature including the param nullability (we're already past the callee). TODO: see if we can update the callee nullability or do this resugaring earlier?

This fixes the `*nullable_vec[i]` case and `**optional`, but not yet the hash_map find `*it->second` case.

PiperOrigin-RevId: 747905063
Change-Id: I961ed4114d59c940927ea6a6a45bbd86634520a9